fix: Object.keys on object with non-configurable props should not fail
Currently it fails with "'getOwnPropertyDescriptor' on proxy: trap reported non-configurability for property 'a' which is either non-existent or configurable in the proxy target"

The reason for that is that the proxy target is the node and not the actual data. See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Proxy/Proxy/getOwnPropertyDescriptor#invariants.